+++ projects/ids-enterprise-typings/lib/toolbar-flex/soho-toolbar-flex.d.ts
@@ -116,6 +116,9 @@ interface SohoToolbarFlexItemStatic {
 
   /** Completely destroys the Toolbar Flex Item component instance. */
   destroy(): void;
+
+  /** Selected anchor in the selected event */
+  selectedAnchor: JQuery;
 }
 
 /**